Use obscure pip-install options to simplify python-casacore install.
We use --global-options to ensure build_ext is performed with
the correct inlude/lib dirs.

Then, we use the -t (target dir) option to pip install to a
simple folder that we can add to PYTHONPATH without futzing
with virtualenvs.

Finally, update the tasks to install to a virtualenv, and make
them optional.